Enhance job synchronization and refactor viewset imports
- Introduced atomic transaction handling in `_sync_jobs_atomically` to ensure all job data is persisted or rolled back on failure, improving data integrity.
- Refactored job synchronization logic in `_collect_data` to utilize the new atomic function, simplifying error handling.
- Updated viewset imports to directly reference `filter_options`, improving code organization and clarity.
