+++ apps/dashboard_reports/awx_queries.py
@@ -36,7 +36,7 @@ def _build_where_clause(join_alias: str, search_str: str | None, pk: Any) -> tup
     if search_str:
         # Escape backslash first, then ILIKE wildcards, so user-supplied % and _ are treated literally.
         escaped = search_str.replace("\\", "\\\\").replace("%", "\\%").replace("_", "\\_")
-        where_clauses.append(f"{join_alias}name ilike %s ESCAPE '\\\\'")
+        where_clauses.append(f"{join_alias}name ilike %s ESCAPE E'\\\\'")
         params.append("%" + escaped + "%")
     if pk is not None:
         where_clauses.append(f"{join_alias}id = %s")
