+++ apps/dashboard_reports/viewsets/job_templates.py
@@ -1,7 +1,7 @@
 """ViewSet for listing AWX job templates as filter dropdown options."""
 
 from apps.dashboard_reports.awx_queries import fetch_templates
-from apps.dashboard_reports.viewsets import FilterOptionsViewSet
+from apps.dashboard_reports.viewsets.filter_options import FilterOptionsViewSet
 
 
 class JobTemplatesViewSet(FilterOptionsViewSet):
