+++ apps/dashboard_reports/viewsets/organizations.py
@@ -1,7 +1,7 @@
 """ViewSet for listing AWX organizations as filter dropdown options."""
 
 from apps.dashboard_reports.awx_queries import fetch_organizations
-from apps.dashboard_reports.viewsets import FilterOptionsViewSet
+from apps.dashboard_reports.viewsets.filter_options import FilterOptionsViewSet
 
 
 class OrganizationsViewSet(FilterOptionsViewSet):
