+++ apps/tasks/v1/views.py
@@ -30,13 +30,13 @@ Operations:
 from datetime import timedelta
 from typing import Any
 
-from ansible_base.rbac.api.permissions import IsSystemAdminOrAuditor
 from django.http import HttpRequest
 from django.utils import timezone
 from rest_framework import status
 from rest_framework.decorators import action
 from rest_framework.response import Response
 
+from apps.core.permissions import DeveloperModeRequired
 from apps.tasks.api_utils import build_error_response
 from apps.tasks.models import Task, TaskExecution
 
@@ -61,7 +61,7 @@ class TaskViewSet(BaseViewSet):
 
     queryset = Task.objects.select_related("created_by").all()
     serializer_class = TaskSerializer
-    permission_classes = [IsSystemAdminOrAuditor]
+    permission_classes = [DeveloperModeRequired]
     ordering_fields = ["id", "name", "status", "scheduled_time", "created", "started_at", "completed_at"]
     ordering = ["-id"]
 
@@ -475,7 +475,7 @@ class TaskExecutionViewSet(BaseViewSet):
 
     queryset = TaskExecution.objects.select_related("task").all()
     serializer_class = TaskExecutionSerializer
-    permission_classes = [IsSystemAdminOrAuditor]
+    permission_classes = [DeveloperModeRequired]
     ordering_fields = ["started_at", "completed_at", "execution_time_seconds", "status"]
     ordering = ["-started_at"]
